services/perplexity-synth-worker/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
pydantic>=2.6
pydantic-settings>=2.2

google-cloud-firestore>=2.13
google-cloud-storage>=2.14
google-cloud-pubsub>=2.19

requests>=2.31
